requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.20.0
mysql-connector-python==8.3.0
statsmodels==0.14.1
